hf_space_deployment/requirements.txt
fastapi
uvicorn
requests
python-multipart
torch
diffusers
transformers
accelerate
pillow
numpy
opencv-python
huggingface_hub
